hello-world/requirements.txt
google-cloud-secret-manager==2.0.0
tweepy
python-dateutil
requests